+++ ConnectExample/ConnectExampleUITests/ConnectExampleUITests.swift
@@ -43,7 +43,7 @@ final class ConnectExampleUITests: XCTestCase {
         
         XCTAssertTrue(loginScreen.awaitState(loginStatus: .passkeyOneTap))
         let profileScreen2 = loginScreen.loginWithOneTap().autoSkip()
-        try waitForCondition(timeout: 10) { profileScreen2.countNumberOfPasskeys() == 1}
+        try waitForCondition() { profileScreen2.countNumberOfPasskeys() == 1}
         
         profileScreen2.appendPasskey()
         XCTAssertEqual(profileScreen2.getErrorMessage(), "You already have a passkey that can be used on this device.")
@@ -82,18 +82,18 @@ final class ConnectExampleUITests: XCTestCase {
         
         
         controlServer.createError = nil
-        try waitForCondition(timeout: 10) { profileScreen.countNumberOfPasskeys() == 0}
+        try waitForCondition() { profileScreen.countNumberOfPasskeys() == 0}
         profileScreen2.appendPasskey()
-        try waitForCondition(timeout: 10) { profileScreen.countNumberOfPasskeys() == 1}
+        try waitForCondition() { profileScreen.countNumberOfPasskeys() == 1}
         let passkeyId = profileScreen2.getPasskeyIds()[0]
         profileScreen2.signOut()
         
         XCTAssertTrue(loginScreen.awaitState(loginStatus: .passkeyOneTap))
         let profileScreen3 = loginScreen.loginWithOneTap().autoSkip()
         
-        try waitForCondition(timeout: 10) { profileScreen3.countNumberOfPasskeys() == 1}
+        try waitForCondition() { profileScreen3.countNumberOfPasskeys() == 1}
         profileScreen3.deletePasskey(passkeyId: passkeyId, complete: true)
-        try waitForCondition(timeout: 10) { profileScreen3.countNumberOfPasskeys() == 0}
+        try waitForCondition() { profileScreen3.countNumberOfPasskeys() == 0}
         let loginScreen2 = profileScreen3.signOut()
         
         XCTAssertTrue(loginScreen2.awaitState(loginStatus: .passkeyTextField))
@@ -119,7 +119,7 @@ final class ConnectExampleUITests: XCTestCase {
         
         loginScreen.switchAccount()
         let profileScreen2 = loginScreen.loginWithIdentifier(email: email).autoSkip()
-        try waitForCondition(timeout: 10) { profileScreen2.countNumberOfPasskeys() == 1}
+        try waitForCondition() { profileScreen2.countNumberOfPasskeys() == 1}
     }
     
     /*
@@ -143,7 +143,7 @@ final class ConnectExampleUITests: XCTestCase {
         controlServer.authorizeError = nil
         
         let profileScreen2 = loginScreen.loginOnPasskeyErrorSoft().autoSkip()
-        try waitForCondition(timeout: 10) { profileScreen2.countNumberOfPasskeys() == 1}
+        try waitForCondition() { profileScreen2.countNumberOfPasskeys() == 1}
     }
     
     /*
@@ -164,7 +164,7 @@ final class ConnectExampleUITests: XCTestCase {
         let loginScreen2 = loginScreen.navigateToSignUp().navigateToLogin()
         
         let profileScreen = loginScreen2.loginWithOverlay()
-        try waitForCondition(timeout: 10) { profileScreen.countNumberOfPasskeys() == 1}
+        try waitForCondition() { profileScreen.countNumberOfPasskeys() == 1}
     }
     
     /* 1.2
@@ -196,14 +196,14 @@ final class ConnectExampleUITests: XCTestCase {
         // append must be skipped automatically
         let tOTPSetupScreen = try await signUpScreen.signUpWithValidData(email: email, phoneNumber: TestDataFactory.phoneNumber, password: TestDataFactory.password).autoSkipAfterSignUp()
         let (profileScreen, setupKey) = try await tOTPSetupScreen.setupTOTP(authenticatorApp: authenticatorApp)
-        XCTAssertTrue(profileScreen.visible(timeout: 10.0), "Profile screen must be visible")
+        XCTAssertTrue(profileScreen.visible(), "Profile screen must be visible")
         XCTAssertFalse(profileScreen.passkeyAppendPossible(), "passkey append button must be hidden")
         let loginScreen2 = profileScreen.signOut()
         
         try loginScreen2.loginWithIdentifierAndPassword(email: email, password: TestDataFactory.password, fallback: true)
         let code = try await authenticatorApp.getCode(setupKey)
         let profileScreen2 = loginScreen2.completeLoginWithTOTP(code: code!).autoSkip()
-        XCTAssertTrue(profileScreen2.visible(timeout: 10.0), "Profile screen must be visible")
+        XCTAssertTrue(profileScreen2.visible(), "Profile screen must be visible")
     }
         
     /* 1.13
@@ -222,7 +222,7 @@ final class ConnectExampleUITests: XCTestCase {
         
         let passkeyId = profileScreen.getPasskeyIds()[0]
         profileScreen.deletePasskey(passkeyId: passkeyId)
-        try waitForCondition(timeout: 10) { profileScreen.countNumberOfPasskeys() == 0}
+        try waitForCondition() { profileScreen.countNumberOfPasskeys() == 0}
                 
         let loginScreen = profileScreen.signOut()
         XCTAssertTrue(loginScreen.awaitState(loginStatus: .fallbackFirst, errorMessage: "You previously deleted this passkey. Use your password to log in instead."))
@@ -257,13 +257,13 @@ final class ConnectExampleUITests: XCTestCase {
         initialScreen.block(blockedUrl: "/connect/login/finish")
         loginScreen.navigateToSignUp().navigateToLogin()                
         loginScreen.loginWithIdentifier(email: email)
-        XCTAssertTrue(initialScreen.awaitState(loginStatus: .fallbackFirst, errorMessage: "Passkey error. Use password to log in.", timeout: 10.0))
+        XCTAssertTrue(initialScreen.awaitState(loginStatus: .fallbackFirst, errorMessage: "Passkey error. Use password to log in."))
         initialScreen.unblock()
         
         // successfull login
         loginScreen.navigateToSignUp().navigateToLogin()
         let profileScreen2 = loginScreen.loginWithIdentifier(email: email).autoSkip()
-        XCTAssertTrue(profileScreen2.visible(timeout: 10.0))
+        XCTAssertTrue(profileScreen2.visible())
     }
     
     func testAppendErrorStatesNetworkBlocking() async throws {
@@ -287,7 +287,7 @@ final class ConnectExampleUITests: XCTestCase {
         initialScreen.block(blockedUrl: "/connect/append/init")
         let postLoginScreen = loginScreen.completeLoginWithTOTP(code: code!)
         let profileScreen2 = postLoginScreen.autoSkip()
-        XCTAssertTrue(profileScreen2.visible(timeout: 10.0))
+        XCTAssertTrue(profileScreen2.visible())
         
         let loginScreen2 = profileScreen2.signOut()
         try loginScreen2.loginWithIdentifierAndPassword(email: email, password: TestDataFactory.password)
@@ -296,7 +296,7 @@ final class ConnectExampleUITests: XCTestCase {
         // block append-start
         initialScreen.block(blockedUrl: "/connect/append/start")
         let profileScreen3 = loginScreen.completeLoginWithTOTP(code: code2!).autoSkip()
-        XCTAssertTrue(profileScreen3.visible(timeout: 10.0))
+        XCTAssertTrue(profileScreen3.visible())
         
         let loginScreen3 = profileScreen3.signOut()
         try loginScreen3.loginWithIdentifierAndPassword(email: email, password: TestDataFactory.password)
@@ -305,7 +305,7 @@ final class ConnectExampleUITests: XCTestCase {
         // block append-finish
         initialScreen.block(blockedUrl: "/connect/append/finish")
         let profileScreen4 = loginScreen.completeLoginWithTOTP(code: code3!).autoSkip()
-        XCTAssertTrue(profileScreen4.visible(timeout: 10.0))        
+        XCTAssertTrue(profileScreen4.visible())        
     }
     
     func testManageErrorStatesNetworkBlocking() async throws {
@@ -317,14 +317,14 @@ final class ConnectExampleUITests: XCTestCase {
             .signUpWithValidData(email: email, phoneNumber: TestDataFactory.phoneNumber, password: TestDataFactory.password)
             .append(expectAutoAppend: true)
         
-        XCTAssertTrue(profileScreen.visible(timeout: 10.0))
+        XCTAssertTrue(profileScreen.visible())
         let passkeyCount = profileScreen.countNumberOfPasskeys()
         XCTAssertEqual(passkeyCount, 1)
         
         // block manage-init
         initialScreen.block(blockedUrl: "/connect/manage/init")
         profileScreen.reloadPage()
-        XCTAssertTrue(profileScreen.visible(timeout: 10.0))
+        XCTAssertTrue(profileScreen.visible())
         
         XCTAssertEqual(profileScreen.getErrorMessage(), "Unable to access passkeys. Check your connection and try again.")
         XCTAssertEqual(profileScreen.getListMessage(), "We were unable to show your list of passkeys due to an error. Try again later.")
@@ -332,7 +332,7 @@ final class ConnectExampleUITests: XCTestCase {
         // block manage-list
         initialScreen.block(blockedUrl: "/connect/manage/list")
         profileScreen.reloadPage()
-        XCTAssertTrue(profileScreen.visible(timeout: 10.0))
+        XCTAssertTrue(profileScreen.visible())
         
         XCTAssertEqual(profileScreen.getErrorMessage(), "Unable to access passkeys. Check your connection and try again.")
         XCTAssertEqual(profileScreen.getListMessage(), "We were unable to show your list of passkeys due to an error. Try again later.")
@@ -340,7 +340,7 @@ final class ConnectExampleUITests: XCTestCase {
         // block append-start
         initialScreen.block(blockedUrl: "/connect/append/start")
         profileScreen.reloadPage()
-        XCTAssertTrue(profileScreen.visible(timeout: 10.0))
+        XCTAssertTrue(profileScreen.visible())
         XCTAssertEqual(profileScreen.countNumberOfPasskeys(), 1)
         profileScreen.appendPasskey()
         
@@ -349,11 +349,11 @@ final class ConnectExampleUITests: XCTestCase {
         
         // block append-finish (we have to delete the passkey first, otherwise we get an excludeCredentials error)
         profileScreen.deletePasskey(passkeyId: profileScreen.getPasskeyIds()[0], complete: true)
-        try waitForCondition(timeout: 10) { profileScreen.countNumberOfPasskeys() == 0}
+        try waitForCondition() { profileScreen.countNumberOfPasskeys() == 0}
         
         initialScreen.block(blockedUrl: "/connect/append/finish")
         profileScreen.reloadPage()
-        XCTAssertTrue(profileScreen.visible(timeout: 10.0))
+        XCTAssertTrue(profileScreen.visible())
         profileScreen.appendPasskey()
         
         XCTAssertEqual(profileScreen.getErrorMessage(), "Passkey creation failed. Please try again later.")
@@ -363,7 +363,7 @@ final class ConnectExampleUITests: XCTestCase {
         initialScreen.block(blockedUrl: "/connect/manage/delete")
         // apppend passkey (to prepare for manage-delete)
         profileScreen.appendPasskey()
-        try waitForCondition(timeout: 10) { profileScreen.countNumberOfPasskeys() == 1}
+        try waitForCondition() { profileScreen.countNumberOfPasskeys() == 1}
         
         profileScreen.deletePasskey(passkeyId: profileScreen.getPasskeyIds()[0], complete: true)
         XCTAssertEqual(profileScreen.getErrorMessage(), "Passkey deletion failed. Please try again later.")