+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2025-07-02
+### Timeout handling
+- Added a default timeout of 10s for all requests going to the Corbado Frontend API
+
+## [1.0.0] - 2025-07-01
+### Make feature complete with @corbado/connect-react (web version of this library) 
+
 ## [0.1.0] - 2025-06-05
 
 ### Added